requirements.txt
EasyProcess==0.2.3
Pillow==5.0.0
PyAutoGUI==0.9.36
PyMsgBox==1.0.6
pyscreenshot==0.4.2
PyScreeze==0.1.13
PyTweening==1.0.3
